feat: componentizing the home and context improvements
