+++ src/views/pages/home/remove-modal.tsx
@@ -1,0 +1,37 @@
+import { Button } from '@/views/components/ui/button'
+
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '../../components/ui/alert-dialog'
+
+interface IRemoveModal {
+  open: boolean
+  onCloseRemoveModal: () => void
+}
+
+export function RemoveModal({ open, onCloseRemoveModal }: IRemoveModal) {
+  return (
+    <AlertDialog open={open}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete contact
+            and remove your data from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <Button variant="outline" onClick={onCloseRemoveModal}>
+            Cancel
+          </Button>
+          <Button variant="destructive">Remove</Button>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
+}
